+++ Ynov.API/DTOitem/CardInputModel.cs
@@ -2,7 +2,7 @@ namespace Ynov.API.DTOitem;
 
 public class CardInputModel
 {
-    public int BoardId { get; set; }
+    public long BoardId { get; set; }
     public string Description { get; set; }
     public string Name { get; set; }
 }
